added custom hook for responsiveness
